tests/render_engine_test.rs
#[cfg(test)]
mod tests {
    use rust_2d_game_engine::render_engine::Renderer;
    use wgpu::ShaderModuleDescriptor;
    use futures::executor::block_on;

    #[test]
    fn test_renderer_initialization() {
        let _renderer = Renderer::new();  // Initialize renderer
        assert!(true, "Renderer initialized successfully");
    }

    #[test]
    fn test_texture_creation() {
        let _renderer = Renderer::new();

        let texture_extent = wgpu::Extent3d {
            width: 1024,
            height: 1024,
            depth_or_array_layers: 1,
        };

        assert_eq!(texture_extent.width, 1024);
        assert_eq!(texture_extent.height, 1024);
        assert_eq!(texture_extent.depth_or_array_layers, 1);
    }

    #[test]
    fn test_render_scene_executes() {
        let mut renderer = Renderer::new();
        renderer.render_scene();
        assert!(true, "render_scene should execute without errors");
    }

    #[test]
    fn test_instance_initialization() {
        let instance_desc = wgpu::InstanceDescriptor {
            backends: wgpu::Backends::all(),
            dx12_shader_compiler: wgpu::Dx12Compiler::Fxc,
            flags: wgpu::InstanceFlags::empty(),
            gles_minor_version: wgpu::Gles3MinorVersion::default(),
        };
        let _instance = wgpu::Instance::new(instance_desc);
        assert!(true, "Instance should be created successfully");
    }

    #[test]
    fn test_request_device() {
        let instance_desc = wgpu::InstanceDescriptor {
            backends: wgpu::Backends::all(),
            dx12_shader_compiler: wgpu::Dx12Compiler::Fxc,
            flags: wgpu::InstanceFlags::empty(),
            gles_minor_version: wgpu::Gles3MinorVersion::default(),
        };
        let instance = wgpu::Instance::new(instance_desc);

        let adapter = block_on(instance.request_adapter(&wgpu::RequestAdapterOptions {
            power_preference: wgpu::PowerPreference::HighPerformance,
            compatible_surface: None,
            force_fallback_adapter: false,
        })).unwrap();

        let (device, _queue) = block_on(adapter.request_device(
            &wgpu::DeviceDescriptor {
                label: None,
                required_features: wgpu::Features::empty(),
                required_limits: wgpu::Limits::default(),
                memory_hints: wgpu::MemoryHints::default(),
            },
            None,
        )).unwrap();

        assert!(device.limits().max_texture_dimension_2d > 0);
    }

    #[test]
    fn test_error_handling_in_renderer() {
        let result = std::panic::catch_unwind(|| {
            let mut renderer = Renderer::new();
            renderer.render_scene();
        });

        assert!(result.is_ok(), "Renderer should not panic when executing render_scene");
    }

    #[test]
    fn test_shader_compilation() {
        let renderer = Renderer::new();
        let shader_source = r#"
            struct VertexOutput {
                @builtin(position) pos: vec4<f32>,
            };

            @vertex
            fn vs_main() -> VertexOutput {
                var out: VertexOutput;
                out.pos = vec4<f32>(0.0, 0.0, 0.0, 1.0);
                return out;
            }

            @fragment
            fn fs_main() -> @location(0) vec4<f32> {
                return vec4<f32>(0.2, 0.3, 0.4, 1.0);
            }
        "#;

        let _shader_module = renderer.device.create_shader_module(ShaderModuleDescriptor {
            label: Some("Test Shader"),
            source: wgpu::ShaderSource::Wgsl(shader_source.into()),
        });

        assert!(true, "Shader compiled successfully");
    }

    #[test]
    fn test_high_load_rendering() {
        let mut renderer = Renderer::new();
        for _ in 0..10000 {
            renderer.render_scene();
        }
        assert!(true, "Engine handled high-load rendering without crashing");
    }
}